requirements.txt
matplotlib~=3.7.1
pandas~=1.5.3
numpy~=1.24.2
python-dotenv~=1.0.0